Removed createMenuBar in MainFrame, renamed and restructured packages
