+++ WordGuess/src/gui/panels/SubmitPanel.java
@@ -1,4 +1,4 @@
-package GUI.InterfacePanels;
+package gui.panels;
 
 import java.awt.Dimension;
 import java.awt.FlowLayout;
