+++ WordGuess/src/gui/panels/ToolbarPanel.java
@@ -1,4 +1,4 @@
-package GUI.InterfacePanels;
+package gui.panels;
 
 import java.awt.FlowLayout;
 import java.awt.event.ActionEvent;
@@ -6,8 +6,6 @@ import java.awt.event.ActionListener;
 
 import javax.swing.BorderFactory;
 import javax.swing.JButton;
-import javax.swing.JFrame;
-import javax.swing.JOptionPane;
 import javax.swing.JPanel;
 
 import tools.Actions;
