+++ WordGuess/src/gui/panels/UserNamePanel.java
@@ -1,10 +1,9 @@
-package GUI.InterfacePanels;
+package gui.panels;
 
 
 import java.awt.Dimension;
 
 import javax.swing.border.Border;
-import javax.swing.table.TableColumn;
 import javax.swing.BorderFactory;
 import javax.swing.JPanel;
 import javax.swing.JScrollPane;
